requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.23
alembic>=1.12.1
pydantic>=2.4.2
python-multipart>=0.0.6
pyaudio>=0.2.13
speechrecognition>=3.10.0
vosk>=0.3.45
langdetect>=1.0.9
opencv-python>=4.8.1
pyautogui>=0.9.54
google-generativeai>=0.3.1
coqui-tts>=0.3.2
transformers>=4.35.0
torch>=2.1.0
websockets>=12.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.3
black>=23.10.1
isort>=5.12.0
mypy>=1.6.1
flake8>=6.1.0